requirements.txt
pandas>=2.2.2
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
plotly>=5.22.0
jupyter>=1.0.0
nbformat>=5.10.4
openpyxl>=3.1.2